soc: riscv: hi3861: Build boot images
This generates loaderboot and flashboot images, which are required for
the entire HiSilicon Hi3861 family.

The loaderboot is an executable image that are loaded to RAM by the
HiBurn runner and interacts with the runner to write the Zephyr kernel
into the flash.

The flashboot is another executable image that loaded by romboot from
flash on each boot. It's only job is to locate and boot up the Zephyr
kernel.

This commit also populates the loaderboot and flashboot paths in the
default runner arguments.

Signed-off-by: Chen Xingyu <hi@xingrz.me>