drivers: intc: Add CONFIG_PLIC_HAS_EDGE_TRIGGER for PLIC
On some platforms, such as T-HEAD C906, edge-triggered interrupts are not
supported, and the `TRIG_TYPE` register does not exist. Accessing such a
register causes mcause 5 (Load access fault).

This commit introduces a Kconfig option to prevent such faults. On
platforms that do not support edge-triggered interrupts, this option should
be set to `n`. For backward compatibility reasons, this option is set to
`y` by default.

Co-authored-by: honglin leng <a909204013@gmail.com>
Signed-off-by: Chen Xingyu <hi@xingrz.me>