samples: boards: milkv_duo: Add sample for mbox driver
This commit adds a mbox sample for the Milk-V Duo, a board equipped with
the SOPHGO CV1800B SoC.

The CV1800B is a multi-core SoC that consists of two C906 cores. The big
core, referred to as C906B, runs Linux, while the little core, referred to
as C906L, runs an RTOS.

This sample receives mailbox requests from the big core running the
official SDK's mailbox sample, prints the request body, and performs
example operations such as LED control.

Signed-off-by: honglin leng <a909204013@gmail.com>